* SQLite support * SQLite dll can be found without installing it in the GAC * SQLite x64 and x86 * x86 and x64 Enhancer support with platform target * Restriction to .NET 4.0 full * Restriction to NDO.dll 2.1
git-svn-id: http://www.netdataobjects.de:81/svn/NDO20/trunk@103 e088091c-91c0-7b4b-9d36-2eafbd79b41f

+++ TestEnhancerVersion4/NdoUnitTests/NDOMitarbeiterTests.cs
@@ -501,9 +501,9 @@ namespace NdoUnitTests
 			}
 			Query q = pm.NewQuery(typeof(Mitarbeiter), null);
 			Mitarbeiter.QueryHelper qh = new Mitarbeiter.QueryHelper();
-			decimal newsum = q.ExecuteAggregate(qh.position.Y, Query.AggregateType.Sum);
+			decimal newsum = (decimal)q.ExecuteAggregate(qh.position.Y, Query.AggregateType.Sum);
 			Assert.AreEqual(sum, newsum, "Summe stimmt nicht");
-			decimal count = q.ExecuteAggregate(qh.position.X, Query.AggregateType.Count);
+			decimal count = (decimal)q.ExecuteAggregate(qh.position.X, Query.AggregateType.Count);
 			Assert.AreEqual(20, count, "Count stimmt nicht");
 		}
